Made some Bug Fixes

+++ routes/userRoutes.js
@@ -1,29 +1,21 @@
 const express = require("express");
-const { getAllUsers, signUp, deleteAll, protectRoute, getUserProfile, login, isAuthorised, revokeAccess } = require("../controllers/userController");
+const { getAllUsers, signUp, deleteAll, protectRoute, getUserProfile, login, isAuthorised, revokeAccess, giveAccess, checkDataAccess } = require("../controllers/userController");
 const router = express.Router();
 const passport = require('passport');
-const { signin, signinCallback, dashboard } = require("../controllers/googleAuthController");
 const downloadData = require("../controllers/nodemailerController");
 
 router.post('/signup', signUp);
 router.post('/login', login);
-router.delete('/delete', deleteAll)
-
-
-
-// Set up Express routes
-router.get('/auth/google', signin);
-router.get('/auth/google/callback', signinCallback, dashboard);
-
-
-
 router.use(protectRoute)
 router
-    .route('/userProfile')
-    .get(getUserProfile)
+.route('/userProfile')
+.get(getUserProfile)
 
+router.use(checkDataAccess)
 router.get('/download', downloadData);
 router.use(isAuthorised(['admin']));
+router.delete('/delete', deleteAll)
 router.get('/allUsers', getAllUsers);
+router.post('/giveAccess', giveAccess)
 router.post('/revokeAccess',revokeAccess)
 module.exports = router;
